hotel的setbestprice

+++ client/src/main/java/businesslogic/bl/availableroombl/HotelInfoAvailService.java
@@ -10,9 +10,9 @@ import util.ResultMessage;
 public interface HotelInfoAvailService {
 	/**
 	 * 可用客房模块设置酒店的最低价格
-	 * @param price int型，hotelID String型，传递酒店最低价格
+	 * @param price double型，hotelID String型，传递酒店最低价格
 	 * @return ResultMessage，设置成功返回SUCCESS,失败返回FAIL，格式错误返回具体什么格式错误
 	 *
 	 */
-     public ResultMessage setBestPrice(int price,String hotelID);
+     public ResultMessage setBestPrice(double price,String hotelID);
 }
